# metadata: Fix bug where AppendToOutgoingContext could modify another context's metadata