Fixes #1925 

For 1-10 append operations, performance is slightly worse using this method, but it is correct and ~3x faster than the old method of copying the whole map.

Before:
```
BenchmarkAppendToOutgoingContext/1-12                          	10000000	       197 ns/op
BenchmarkAppendToOutgoingContext/10-12                         	 1000000	      2033 ns/op
BenchmarkAppendToOutgoingContext/100-12                        	  100000	     15497 ns/op
```

After:
```
BenchmarkAppendToOutgoingContext/1-12                          	10000000	       193 ns/op
BenchmarkAppendToOutgoingContext/10-12                         	  500000	      2513 ns/op
BenchmarkAppendToOutgoingContext/100-12                        	   30000	     51658 ns/op
```

The old method for reference:
```
Benchmark_AddingMetadata_ContextManipulationApproach/1-12  	 2000000	       805 ns/op
Benchmark_AddingMetadata_ContextManipulationApproach/10-12 	  200000	      7925 ns/op
Benchmark_AddingMetadata_ContextManipulationApproach/100-12    20000	     80581 ns/op
```

(Note: I'm also copying `kv...`, because it could be passed as `mySlice...`, and I also added a test to demonstrate/catch that.)